+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/java/LambdaUtils.java
@@ -122,7 +122,8 @@ public final class LambdaUtils {
 
     public static boolean isLambdaType(ResolvedJavaType type) {
         String typeName = type.getName();
-        return type.isFinalFlagSet() && isLambdaName(typeName);
+        return !isArrayTypeName(typeName) && !type.isArray() &&
+                        type.isFinalFlagSet() && isLambdaName(typeName);
     }
 
     public static boolean isLambdaName(String name) {
@@ -196,23 +197,13 @@ public final class LambdaUtils {
     }
 
     /**
-     * Checks if the passed class is a lambda class.
-     *
-     * @param type the type to be checked
-     * @return true if type is a lambda class, false instead
-     */
-    public static boolean isLambdaClass(ResolvedJavaType type) {
-        return isLambdaClassName(type.toClassName());
-    }
-
-    /**
      * Checks if the passed class is lambda class.
      *
      * @param clazz class to be checked
      * @return true if the clazz is lambda class, false instead
      */
     public static boolean isLambdaClass(Class<?> clazz) {
-        return isLambdaClassName(clazz.getName());
+        return !clazz.isArray() && isLambdaClassName(clazz.getName());
     }
 
     /**
@@ -223,5 +214,9 @@ public final class LambdaUtils {
      */
     public static boolean isLambdaClassName(String className) {
         return className.contains(LAMBDA_CLASS_NAME_SUBSTRING);
+    }
+
+    private static boolean isArrayTypeName(String name) {
+        return !name.isEmpty() && name.charAt(0) == '[';
     }
 }